requirements.txt
feedparser==6.0.10
newspaper3k==0.2.8
apache-airflow-providers-telegram==4.1.1